pyproject.toml
[project]
name = "sesai-platform"
version = "0.1.0"
description = "Enterprise-grade Turkish Voice AI Platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SesAI Team"}
]

dependencies = [
    # Core Voice Pipeline
    "pipecat-ai[silero]>=0.0.50",
    "faster-whisper>=1.0.0",
    "chatterbox-tts>=0.1.0",

    # API Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # Audio Processing
    "numpy>=1.26.0,<2.4",
    "scipy>=1.12.0",
    "soundfile>=0.12.0",

    # LLM Client
    "openai>=1.12.0",

    # Utilities
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
sesai = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
